requirements.txt
aiofiles==23.2.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Faker==20.1.0
fastapi==0.104.1
gmssl==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.2
macholib==1.16.3
MarkupSafe==3.0.2
mysql-connector-python==8.2.0
packaging==25.0
protobuf==4.21.12
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
